+++ CHANGELOG.md
@@ -6,6 +6,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 * Fix `flexberry-lookup` readonly-mode when autocomplete is enabled.
 
+### Added
+* Add option `showValidationMessagesInRow` for `object-list-view`, `flexberry-objectlistview` & `flexberry-groupedit` components.
+  Option is `false` by default in `object-list-view` & `flexberry-objectlistview`, but is `true` by default in `flexberry-groupedit`.
+  If `true` then validation message will be shown in each cell containing invalid value.
+
 ## [0.8.0-beta.17] - 2016-12-08
 ### Fixed
 * Fix `perPage` settings in modal.