+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+* Fix `filterCondition` when refreshing page.
 
 ## [0.8.0-beta.18] - 2016-12-12
 ### Added